requirements.txt
streamlit==1.31.1
python-docx==0.8.11
PyMuPDF==1.23.8
pytesseract==0.3.10
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
spacy==3.7.2
textstat==0.7.3
scikit-learn==1.4.0
torch==2.1.2
numpy==1.24.3
pandas==2.0.3
openai>=1.3.0
